Changed card stacks to template method
Templated used
